Master - CRUD methods fixed, User adding at /start implemented

+++ orm/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, BigInteger
 from sqlalchemy.orm import relationship
 
 from .database import Base
@@ -7,7 +7,7 @@ from .database import Base
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     name = Column(String, unique=False)
     # email = Column(String, unique=True, index=True)
     # items = relationship("Item", back_populates="owner")
